+++ icekit/page_types/author/migrations/0011_auto_20170518_1629.py
@@ -1,0 +1,34 @@
+# -*- coding: utf-8 -*-
+from __future__ import unicode_literals
+
+from django.db import migrations, models
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        ('icekit_authors', '0010_auto_20170317_1656'),
+    ]
+
+    operations = [
+        migrations.AddField(
+            model_name='author',
+            name='admin_notes',
+            field=models.TextField(help_text=b"Administrator's notes about this content", blank=True),
+        ),
+        migrations.AddField(
+            model_name='author',
+            name='brief',
+            field=models.TextField(help_text=b'A document brief describing the purpose of this content', blank=True),
+        ),
+        migrations.AddField(
+            model_name='authorlisting',
+            name='admin_notes',
+            field=models.TextField(help_text=b"Administrator's notes about this content", blank=True),
+        ),
+        migrations.AddField(
+            model_name='authorlisting',
+            name='brief',
+            field=models.TextField(help_text=b'A document brief describing the purpose of this content', blank=True),
+        ),
+    ]