+++ icekit/page_types/search_page/admin.py
@@ -1,10 +1,7 @@
-from fluent_pages.integration.fluent_contents.admin import FluentContentsPageAdmin
+from fluent_pages.integration.fluent_contents.admin \
+    import FluentContentsPageAdmin
 
 from icekit.admin import ICEkitContentsAdmin
-
-
-class UnpublishableSearchPageAdmin(FluentContentsPageAdmin):
-    placeholder_layout_template = 'icekit/page_types/search_page/default.html'
 
 
 class SearchPageAdmin(FluentContentsPageAdmin, ICEkitContentsAdmin):