+++ docs/changelog.rst
@@ -1,8 +1,21 @@
 Changelog
 =========
 
+In development
+--------------
+
+-  Updated glamkit-collections to use new EDTF fields and automatically derive
+   EDTF date from plain text date.
+
+-  Remove unpublishable variants of the main GLAMkit models as keeping them is
+   not worth the maintenance overhead or chance for confusion.
+
 0.17 (2017-04-30)
 -----------------
+
+-  Rename the GET parameter used to trigger and permit preview of draft pages
+   to be 'preview' instead of 'edit'. The legacy 'edit' parameter name is still
+   supported.
 
 -  Changed the way that we specify the homepage: create a page with url
    override `/`. This is the way fluent does it. As a fallback, we look for a
@@ -48,6 +61,9 @@ Changelog
       The existing ``ThumbnailAdminMixin`` now inherits from this class and
       consequently, models which use ``ThumbnailAdminMixin`` have a visual
       preview in ``raw_id`` fields, including ``icekit.plugins.image.Image``.
+
+   - The "icekit/admin/fluent_layouts_change_form.html" template has moved to
+      "admin/fluent_layouts_change_form.html". Please update your references.
 
 -  Add ``icekit.api`` application to provide APIs to read data from, and
    sometimes write data to, items in the site from program clients.